\name{ReadGenea-package}
\alias{ReadGenea-package}
\alias{ReadGenea}
\docType{package}
\title{
ReadGenea: a package to process binary accelerometer output files.
}
\description{
This is a package to process binary output files from the Genea accelerometer data.  The main function is:\cr\cr

read.bin\cr
}
\details{
\tabular{ll}{
Package: \tab ReadGenea\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 12/06/2012\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Zhou Fang <zhou@activinsights.co.uk>
ActivInsights Ltd. <joss.langford@activinsights.co.uk>
}
\keyword{ package }
